Import of AMBS/Lingua-Jspell-1.81 from CPAN.
gitpan-cpan-distribution: Lingua-Jspell
gitpan-cpan-version:      1.81
gitpan-cpan-path:         AMBS/Lingua/Lingua-Jspell-1.81.tar.gz
gitpan-cpan-author:       AMBS
gitpan-cpan-maturity:     released

+++ src/lookup.c
@@ -266,15 +266,15 @@ static int act_all_entry(void)
 
    for (i = numsflags + numpflags, entry = sflaglist; --i >= 0; entry++) {
       if (entry->stripl)
-         entry->strip = (ichar_t *) &hashstrings[(int) entry->strip];
+         entry->strip = (ichar_t *) &hashstrings[(long int) entry->strip];
       else
          entry->strip = NULL;
       if (entry->affl)
-         entry->affix = (ichar_t *) &hashstrings[(int) entry->affix];
+         entry->affix = (ichar_t *) &hashstrings[(long int) entry->affix];
       else
          entry->affix = NULL;
       if (entry->classl)
-         entry->jclass = (ichar_t *) &hashstrings[(int) entry->jclass];
+         entry->jclass = (ichar_t *) &hashstrings[(long int) entry->jclass];
       else
          entry->jclass = NULL;
    }
